% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.alexa.package.R
\name{alexa_check}
\alias{alexa_check}
\title{Request Response Verification}
\usage{
alexa_check(req)
}
\arguments{
\item{req}{request}
}
\value{
in case of failure, a message
}
\description{
Request Response Verification
}
