% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.alexa.package.R
\name{alexa_GET}
\alias{alexa_GET}
\title{Base POST AND GET functions. Not exported.}
\usage{
alexa_GET(query, key = Sys.getenv("AWS_ACCESS_KEY_ID"),
  secret = Sys.getenv("AWS_SECRET_ACCESS_KEY"), ...)
}
\arguments{
\item{query}{query list}

\item{key}{A character string containing an AWS Access Key ID. The default is retrieved from \code{Sys.getenv("AWS_ACCESS_KEY_ID")}.}

\item{secret}{A character string containing an AWS Secret Access Key. The default is retrieved from \code{Sys.getenv("AWS_SECRET_ACCESS_KEY")}.}

\item{\dots}{Additional arguments passed to \code{\link[httr]{GET}}.}
}
\value{
list
}
\description{
GET
}
