% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.alexa.package.R
\name{alexa_PROCESS}
\alias{alexa_PROCESS}
\title{Postprocess the results a bit}
\usage{
alexa_PROCESS(res)
}
\arguments{
\item{res}{result}
}
\value{
display request ID and Response Status and the first member of the list
}
\description{
Postprocess the results a bit
}
