% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.alexa.package.R
\docType{package}
\name{aws.alexa-package}
\alias{aws.alexa}
\alias{aws.alexa-package}
\title{aws.alexa: R Client for the Alexa Amazon Web Information Service API}
\description{
OCR images, barcodes, forms, documents with machine readable zones, e.g. passports, right from R.
Get the results in any one of a wide variety of formats, from text to XML. 
The package provides access to Abbyy Cloud OCR. For more information, see \url{http://ocrsdk.com/}. 
Details about results of calls to the API can be found at \url{http://docs.aws.amazon.com/AlexaWebInfoService/latest/}.

To learn how to use aws.alexa, see this vignette: \url{vignettes/overview.html}. 

You need to get credentials (Access Key ID and Secret Access Key) to use this application. 
If you haven't already, get these at \url{https://console.aws.amazon.com/iam/home#security_credential}. 
And set these using \code{\link{set_secret_key}}
}
\author{
Gaurav Sood
}

