% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_links.R
\name{in_links}
\alias{in_links}
\title{Sites linking to the site}
\usage{
in_links(url = NULL, start = 0, count = 20, ...)
}
\arguments{
\item{url}{String; Required; valid url}

\item{start}{index of result at which to start; default = 0}

\item{count}{Number of results to return for this request; Max = 20; Default = 20}

\item{\dots}{Additional arguments passed to \code{\link{alexa_GET}}.}
}
\value{
data.frame with two columns: title (site hostname) and url (specific url)
}
\description{
Sites linking to the site
}
\examples{
\dontrun{
in_links(url = "google.com")
}
}
\references{
\url{http://docs.aws.amazon.com/AlexaWebInfoService/latest/ApiReference_SitesLinkingInAction.html}
}
