% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_info.R
\name{url_info}
\alias{url_info}
\title{Get Information about a URL}
\usage{
url_info(url = NULL, response_group = "SiteData")
}
\arguments{
\item{url}{String; Required; valid url}

\item{response_group}{String; Required; One of the following: RelatedLinks, Categories, Rank,
RankByCountry, UsageStats, AdultContent, Speed, Language,  OwnedDomains, LinksInCount, SiteData}
}
\value{
data.frame
}
\description{
Get Information about a URL
}
\examples{
\dontrun{
url_info(domain="http://www.google.com")
}
}
\references{
\url{http://docs.aws.amazon.com/AlexaWebInfoService/latest/ApiReference_UrlInfoAction.html}
}

