% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_info.R
\name{url_info}
\alias{url_info}
\title{Get Information about a URL}
\usage{
url_info(url = NULL, response_group = "SiteData", ...)
}
\arguments{
\item{url}{String; Required; valid url}

\item{response_group}{String; Required; One of the following: RelatedLinks, Categories, Rank,
RankByCountry, UsageStats, AdultContent, Speed, Language,  OwnedDomains, LinksInCount, SiteData
Default is 'SiteData'. Multiple fields can be passed. They must be separated by comma.}

\item{\dots}{Additional arguments passed to \code{\link{alexa_GET}}.}
}
\value{
named list
}
\description{
Get Information about a URL
}
\examples{
\dontrun{
url_info(url = "http://www.google.com")
}
}
\references{
\url{http://docs.aws.amazon.com/AlexaWebInfoService/latest/ApiReference_UrlInfoAction.html}
}
